Moved sidebar and its children into a Sidebar directory to keep everything organised
