+++ components/Sidebar/Sidebar.tsx
@@ -4,9 +4,9 @@ import React, { useMemo } from "react";
 import { usePathname } from "next/navigation";
 import { BiSearch, BiHomeAlt2 } from "react-icons/bi";
 // import { HiHome } from "react-icons/hi";
-import Box from "./Box";
+import Box from "../Box";
 import SidebarItem from "./SidebarItem";
-import Library from "./Library";
+import Library from "../Library";
 import { Song } from "@/types/types";
 import usePlayer from "@/hooks/usePlayer";
 import { twMerge } from "tailwind-merge";
